refactor: 댓글 코드 분리
