+++ week07/Gangbean/src/apis/lp.ts
@@ -2,7 +2,6 @@ import type { PaginationDto } from '../types/common';
 import type {
     RequestCreateLpDto,
     ResponseCreateLpDto,
-    ResponseLpCommentListDto,
     ResponseLpDetailDto,
     ResponseLpListDto,
 } from '../types/lp';
@@ -23,20 +22,6 @@ export const getLpDetail = async (
 ): Promise<ResponseLpDetailDto> => {
     const { data } = await axiosInstance.get(
         `/v1/lps/${lpId}`
-    );
-
-    return data;
-};
-
-export const getLpComments = async (
-    lpId: number,
-    paginationDto: PaginationDto
-): Promise<ResponseLpCommentListDto> => {
-    const { data } = await axiosInstance.get(
-        `/v1/lps/${lpId}/comments`,
-        {
-            params: paginationDto,
-        }
     );
 
     return data;
