+++ week07/Gangbean/src/types/comment.ts
@@ -1,0 +1,16 @@
+import type { CursorBasedResponse } from './common';
+import type { Author } from './lp';
+
+export type LpComment = {
+    id: number;
+    content: string;
+    lpId: number;
+    authorId: number;
+    createdAt: Date;
+    updatedAt: Date;
+    author: Author;
+};
+
+export type ResponseLpCommentListDto = CursorBasedResponse<
+    LpComment[]
+>;
